wordcloud_demo/requirements.txt
alabaster==0.7.10
astroid==1.5.3
attrs==17.3.0
Babel==2.5.1
bleach==2.1.2
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cycler==0.10.0
decorator==4.1.2
Django==1.10.6
django-crispy-forms==1.6.1
django-filter==1.0.4
django-formtools==2.0
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
docutils==0.14
entrypoints==0.2.3
Flask==0.12.2
future==0.16.0
html5lib==1.0.1
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.13
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.2.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
lxml==4.1.1
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.1.1
mccabe==0.6.1
mistune==0.8.3
mysqlclient==1.3.10
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
numpy==1.13.3
numpydoc==0.7.0
olefile==0.44
pandocfilters==1.4.2
parso==0.1.1
pickleshare==0.7.4
Pillow==4.1.1
pluggy==0.6.0
prompt-toolkit==1.0.15
psutil==5.4.2
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.1
PyMySQL==0.7.11
pyparsing==2.2.0
pytest==3.3.1
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
pyzmq==16.0.3
qiniu==7.2.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
rope==0.10.7
scikit-image==0.13.1
scipy==1.0.0
selenium==3.8.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
win-unicode-console==0.5
wordcloud==1.3.2
wrapt==1.10.10
